Remove InTransfer/OutTransfer implementation from Lisk Elements
